refactor(bayn): isolate evidence and ledger cores
