--- services/bayn/src/audit/reference/index.ts
@@ -1,8 +1,0 @@
-export { evaluateReference, measureReferenceEvaluationWork } from './evaluation'
-export type {
-  ReferenceEvaluation,
-  ReferenceEvaluationFailure,
-  ReferenceEvaluationWork,
-  ReferenceReplayWork,
-} from './model'
-export { restrictReferenceBuyFill } from './replay'
