+++ services/bayn/src/audit/reference/replay.ts
@@ -1,25 +1,18 @@
-import { pipe, Result } from 'effect'
+import { Result } from 'effect'
 
 import {
   accrueCashYield,
   calculateSessionFees,
-  desiredQuantityMicros,
   elapsedCalendarDays,
   makeFillTerms,
-  makeOrderOutcome,
-  microsToNumber,
   notionalMicros,
   ppm,
-  referencePriceMicros,
   saleCostBasisMicros,
   scaleQuantityMicros,
-  type FeeInput,
-  type FillTerms,
 } from '../../execution-model'
 import { canonicalHashV1 } from '../../hash'
 import type {
   CashChange,
-  DailyBar,
   DailyPerformancePoint,
   DailyPositionMark,
   DecisionEvent,
@@ -27,309 +20,24 @@ import type {
   FeeEvent,
   FillEvent,
   IsoDate,
-  PerformanceMetrics,
   SignalDecision,
   SimulatedOrder,
   SimulationProtocol,
 } from '../../types'
-import { average, sampleDeviation, tradingDays } from './decisions'
 import type { Position, ReferenceComputation, ReplayWithWork, Session, Target } from './model'
+import {
+  calculateReplayMetrics,
+  makeCashChange,
+  makeReferenceFill,
+  makeReferenceOrder,
+  replayBuysFitCash,
+  replayDesiredQuantities,
+  replayPositionValue,
+  replayPrices,
+  restrictReferenceBuyFill,
+} from './replay/support'
 
-const metrics = (
-  equityMicros: readonly bigint[],
-  turnoverMicros: bigint,
-  feeMicros: bigint,
-  spreadMicros: bigint,
-  slippageMicros: bigint,
-  yieldMicros: bigint,
-  initialMicros: bigint,
-): ReferenceComputation<PerformanceMetrics> => {
-  const firstNonPositiveIndex = equityMicros.findIndex((value) => value <= 0n)
-  if (equityMicros.length < 2 || firstNonPositiveIndex !== -1) {
-    return Result.fail({
-      _tag: 'ReferenceInvalidEquityCurve',
-      observationCount: equityMicros.length,
-      firstNonPositiveIndex: firstNonPositiveIndex === -1 ? null : firstNonPositiveIndex,
-      firstNonPositiveValueMicros:
-        firstNonPositiveIndex === -1 ? null : (equityMicros[firstNonPositiveIndex]?.toString() ?? null),
-    })
-  }
-  const endingEquityMicros = equityMicros.at(-1)
-  if (endingEquityMicros === undefined) {
-    return Result.fail({
-      _tag: 'ReferenceInvalidEquityCurve',
-      observationCount: equityMicros.length,
-      firstNonPositiveIndex: null,
-      firstNonPositiveValueMicros: null,
-    })
-  }
-  const equity = equityMicros.map(microsToNumber)
-  const initial = microsToNumber(initialMicros)
-  const endingEquity = microsToNumber(endingEquityMicros)
-  const returns = equity.map((value, index) => value / (index === 0 ? initial : equity[index - 1]) - 1)
-  const totalReturn = endingEquity / initial - 1
-  const annualizedReturn = Math.pow(endingEquity / initial, tradingDays / equity.length) - 1
-  const annualizedVolatility = sampleDeviation(returns) * Math.sqrt(tradingDays)
-  const sharpe = annualizedVolatility === 0 ? 0 : (average(returns) * tradingDays) / annualizedVolatility
-  let peak = initial
-  let maximumDrawdown = 0
-  for (const value of equity) {
-    peak = Math.max(peak, value)
-    maximumDrawdown = Math.max(maximumDrawdown, 1 - value / peak)
-  }
-  return Result.succeed({
-    observations: equity.length,
-    totalReturn,
-    annualizedReturn,
-    annualizedVolatility,
-    sharpe,
-    maximumDrawdown,
-    annualTurnover: microsToNumber(turnoverMicros) / initial / (equity.length / tradingDays),
-    totalFeesMicros: feeMicros.toString(),
-    totalSpreadCostMicros: spreadMicros.toString(),
-    totalSlippageCostMicros: slippageMicros.toString(),
-    totalCashYieldMicros: yieldMicros.toString(),
-    endingEquityMicros: endingEquityMicros.toString(),
-  })
-}
-
-const order = (
-  runId: string,
-  decision: DecisionEvent,
-  sessionDate: IsoDate,
-  symbol: string,
-  side: 'buy' | 'sell',
-  requestedQuantityMicros: bigint,
-  referencePrice: bigint,
-  protocol: SimulationProtocol,
-): ReferenceComputation<SimulatedOrder> =>
-  pipe(
-    makeOrderOutcome({
-      identity: {
-        schemaVersion: 'bayn.partial-fill-seed.v1',
-        signalDate: decision.signalDate,
-        executionDate: decision.executionDate,
-        symbol,
-        side,
-      },
-      side,
-      requestedQuantityMicros,
-      referencePriceMicros: referencePrice,
-      model: protocol.executionModel,
-    }),
-    Result.map((outcome) => {
-      const material = {
-        decisionId: decision.id,
-        sessionDate,
-        symbol,
-        side,
-        requestedQuantityMicros: outcome.requestedQuantityMicros.toString(),
-        filledQuantityMicros: outcome.filledQuantityMicros.toString(),
-        status: outcome.status,
-        rejectionReason: outcome.rejectionReason,
-        unfilledRemainder: outcome.unfilledRemainder,
-      }
-      return { id: canonicalHashV1({ runId, kind: 'order', ...material }), ...material }
-    }),
-  )
-
-export const restrictReferenceBuyFill = (
-  runId: string,
-  simulatedOrder: SimulatedOrder,
-  permittedQuantity: bigint,
-): ReferenceComputation<SimulatedOrder> => {
-  const modeledQuantity = BigInt(simulatedOrder.filledQuantityMicros)
-  if (modeledQuantity === 0n || modeledQuantity === permittedQuantity) return Result.succeed(simulatedOrder)
-  if (permittedQuantity < 0n || permittedQuantity > modeledQuantity) {
-    return Result.fail({
-      _tag: 'ReferenceBuyFillRestrictionInvalid',
-      orderId: simulatedOrder.id,
-      modeledQuantityMicros: modeledQuantity.toString(),
-      permittedQuantityMicros: permittedQuantity.toString(),
-    })
-  }
-  const material = {
-    decisionId: simulatedOrder.decisionId,
-    sessionDate: simulatedOrder.sessionDate,
-    symbol: simulatedOrder.symbol,
-    side: simulatedOrder.side,
-    requestedQuantityMicros: simulatedOrder.requestedQuantityMicros,
-    filledQuantityMicros: permittedQuantity.toString(),
-    status: permittedQuantity === 0n ? ('rejected' as const) : ('partially-filled' as const),
-    rejectionReason: permittedQuantity === 0n ? ('insufficient-buying-power' as const) : null,
-    unfilledRemainder: 'canceled' as const,
-  }
-  return Result.succeed({ id: canonicalHashV1({ runId, kind: 'order', ...material }), ...material })
-}
-
-const fill = (
-  runId: string,
-  decision: DecisionEvent,
-  simulatedOrder: SimulatedOrder,
-  terms: FillTerms,
-  costBasisMicros: bigint,
-): FillEvent => {
-  const material = {
-    orderId: simulatedOrder.id,
-    decisionId: decision.id,
-    sessionDate: simulatedOrder.sessionDate,
-    symbol: simulatedOrder.symbol,
-    side: simulatedOrder.side,
-    quantityMicros: simulatedOrder.filledQuantityMicros,
-    referencePriceMicros: terms.referencePriceMicros.toString(),
-    priceMicros: terms.fillPriceMicros.toString(),
-    notionalMicros: terms.notionalMicros.toString(),
-    spreadCostMicros: terms.spreadCostMicros.toString(),
-    slippageCostMicros: terms.slippageCostMicros.toString(),
-    costBasisMicros: costBasisMicros.toString(),
-  }
-  return { kind: 'fill', id: canonicalHashV1({ runId, kind: 'fill', ...material }), ...material }
-}
-
-const cashChange = (
-  runId: string,
-  source:
-    | Pick<FillEvent | FeeEvent, 'kind' | 'id' | 'sessionDate'>
-    | { kind: 'cash-yield'; id: string; sessionDate: IsoDate },
-  amountMicros: bigint,
-  cashAfterMicros: bigint,
-): CashChange => {
-  const material = {
-    sourceKind: source.kind,
-    sourceId: source.id,
-    sessionDate: source.sessionDate,
-    amountMicros: amountMicros.toString(),
-    cashAfterMicros: cashAfterMicros.toString(),
-  }
-  return { id: canonicalHashV1({ runId, kind: 'cash-change', ...material }), ...material }
-}
-
-const replayPrices = (
-  session: Session,
-  protocol: SimulationProtocol,
-  price: (bar: DailyBar) => number,
-): ReferenceComputation<Readonly<Record<string, bigint>>> =>
-  pipe(
-    Result.all(
-      protocol.universe.map((symbol) =>
-        pipe(
-          referencePriceMicros(price(session.bars[symbol]), protocol.executionModel),
-          Result.map((priceMicros) => [symbol, priceMicros] as const),
-        ),
-      ),
-    ),
-    Result.map((entries) => Object.fromEntries(entries)),
-  )
-
-const replayPositionValue = (
-  prices: Readonly<Record<string, bigint>>,
-  positions: ReadonlyMap<string, Position>,
-  protocol: SimulationProtocol,
-): ReferenceComputation<bigint> =>
-  protocol.universe.reduce<ReferenceComputation<bigint>>(
-    (total, symbol) =>
-      pipe(
-        total,
-        Result.flatMap((value) =>
-          pipe(
-            notionalMicros(positions.get(symbol)?.quantityMicros ?? 0n, prices[symbol]),
-            Result.map((notional) => value + notional),
-          ),
-        ),
-      ),
-    Result.succeed(0n),
-  )
-
-const replayDesiredQuantities = (
-  equityMicros: bigint,
-  weights: Readonly<Record<string, number>>,
-  prices: Readonly<Record<string, bigint>>,
-  protocol: SimulationProtocol,
-): ReferenceComputation<Readonly<Record<string, bigint>>> =>
-  pipe(
-    Result.all(
-      protocol.universe.map((symbol) =>
-        pipe(
-          desiredQuantityMicros(equityMicros, weights[symbol], prices[symbol], protocol.executionModel),
-          Result.map((quantityMicros) => [symbol, quantityMicros] as const),
-        ),
-      ),
-    ),
-    Result.map((entries) => Object.fromEntries(entries)),
-  )
-
-interface ReferenceBuyCandidate {
-  readonly symbol: string
-  readonly quantityMicros: bigint
-}
-
-const replayBuyFeeInputs = (
-  buys: readonly ReferenceBuyCandidate[],
-  scalePpm: bigint,
-  prices: Readonly<Record<string, bigint>>,
-  protocol: SimulationProtocol,
-  costMultiplierMicros: bigint,
-  minimumNotionalMicros?: bigint,
-): ReferenceComputation<readonly FeeInput[]> =>
-  buys.reduce<ReferenceComputation<readonly FeeInput[]>>(
-    (result, buy) =>
-      pipe(
-        result,
-        Result.flatMap((inputs) =>
-          pipe(
-            scaleQuantityMicros(buy.quantityMicros, scalePpm, protocol.executionModel),
-            Result.flatMap((quantityMicros) => {
-              if (quantityMicros === 0n) return Result.succeed(inputs)
-              return pipe(
-                notionalMicros(quantityMicros, prices[buy.symbol]),
-                Result.flatMap((referenceNotionalMicros) => {
-                  if (minimumNotionalMicros !== undefined && referenceNotionalMicros < minimumNotionalMicros) {
-                    return Result.succeed(inputs)
-                  }
-                  return pipe(
-                    makeFillTerms(
-                      'buy',
-                      quantityMicros,
-                      prices[buy.symbol],
-                      protocol.executionModel,
-                      costMultiplierMicros,
-                    ),
-                    Result.map((terms) => [
-                      ...inputs,
-                      { side: 'buy' as const, quantityMicros, notionalMicros: terms.notionalMicros },
-                    ]),
-                  )
-                }),
-              )
-            }),
-          ),
-        ),
-      ),
-    Result.succeed([]),
-  )
-
-const replayBuysFitCash = (
-  buys: readonly ReferenceBuyCandidate[],
-  scalePpm: bigint,
-  prices: Readonly<Record<string, bigint>>,
-  protocol: SimulationProtocol,
-  costMultiplierMicros: bigint,
-  availableCashMicros: bigint,
-  minimumNotionalMicros?: bigint,
-): ReferenceComputation<boolean> =>
-  pipe(
-    replayBuyFeeInputs(buys, scalePpm, prices, protocol, costMultiplierMicros, minimumNotionalMicros),
-    Result.flatMap((inputs) =>
-      pipe(
-        calculateSessionFees(inputs, protocol.executionModel, costMultiplierMicros),
-        Result.map(
-          (fees) =>
-            inputs.reduce((total, candidate) => total + candidate.notionalMicros, 0n) + fees.totalMicros <=
-            availableCashMicros,
-        ),
-      ),
-    ),
-  )
+export { restrictReferenceBuyFill } from './replay/support'
 
 interface ReplayState {
   readonly positions: ReadonlyMap<string, Position>
@@ -433,7 +141,7 @@ export const replay = (
             ...material,
           }
           events.push(event)
-          changes.push(cashChange(runId, event, accrued, cash))
+          changes.push(makeCashChange(runId, event, accrued, cash))
         }
       }
     }
@@ -519,7 +227,7 @@ export const replay = (
       }
       const sellOrdersResult = Result.all(
         sellPlans.map((candidate) =>
-          order(
+          makeReferenceOrder(
             runId,
             decision,
             session.date,
@@ -541,7 +249,7 @@ export const replay = (
         const requestedNotional = notionalMicros(requested.success, planPrices[candidate.symbol])
         if (Result.isFailure(requestedNotional)) return Result.fail(requestedNotional.failure)
         if (requestedNotional.success < minimumBuyNotionalMicros) continue
-        const simulatedOrder = order(
+        const simulatedOrder = makeReferenceOrder(
           runId,
           decision,
           session.date,
@@ -604,7 +312,7 @@ export const replay = (
         const costBasisResult = saleCostBasisMicros(position.costBasisMicros, quantity, position.quantityMicros)
         if (Result.isFailure(costBasisResult)) return Result.fail(costBasisResult.failure)
         const costBasis = costBasisResult.success
-        const event = fill(runId, decision, simulatedOrder, terms, costBasis)
+        const event = makeReferenceFill(runId, decision, simulatedOrder, terms, costBasis)
         cash += terms.notionalMicros
         turnover += terms.notionalMicros
         spread += terms.spreadCostMicros
@@ -616,7 +324,7 @@ export const replay = (
         sessionFills.push(event)
         if (retainTrace) {
           events.push(event)
-          changes.push(cashChange(runId, event, terms.notionalMicros, cash))
+          changes.push(makeCashChange(runId, event, terms.notionalMicros, cash))
         }
       }
 
@@ -633,7 +341,7 @@ export const replay = (
         )
         if (Result.isFailure(termsResult)) return Result.fail(termsResult.failure)
         const terms = termsResult.success
-        const event = fill(runId, decision, simulatedOrder, terms, terms.notionalMicros)
+        const event = makeReferenceFill(runId, decision, simulatedOrder, terms, terms.notionalMicros)
         cash -= terms.notionalMicros
         turnover += terms.notionalMicros
         spread += terms.spreadCostMicros
@@ -646,7 +354,7 @@ export const replay = (
         sessionFills.push(event)
         if (retainTrace) {
           events.push(event)
-          changes.push(cashChange(runId, event, -terms.notionalMicros, cash))
+          changes.push(makeCashChange(runId, event, -terms.notionalMicros, cash))
         }
       }
 
@@ -679,7 +387,7 @@ export const replay = (
             ...material,
           }
           events.push(event)
-          changes.push(cashChange(runId, event, -fee.totalMicros, cash))
+          changes.push(makeCashChange(runId, event, -fee.totalMicros, cash))
         }
       }
       if (cash < 0n) {
@@ -727,11 +435,7 @@ export const replay = (
           marketValueMicros: marketValue.success.toString(),
         })
       }
-      marks.push({
-        ...point,
-        cashMicros: cash.toString(),
-        positions: markedPositions,
-      })
+      marks.push({ ...point, cashMicros: cash.toString(), positions: markedPositions })
     }
     state = {
       positions,
@@ -747,7 +451,7 @@ export const replay = (
     }
   }
 
-  const metricsResult = metrics(
+  const metricsResult = calculateReplayMetrics(
     equity,
     state.turnoverMicros,
     state.feeMicros,
@@ -772,10 +476,6 @@ export const replay = (
           dailyMarks: marks,
         }
       : null,
-    work: {
-      sessionsProcessed: daily.length,
-      positionStateCopies,
-      positionWrites,
-    },
+    work: { sessionsProcessed: daily.length, positionStateCopies, positionWrites },
   })
 }
