+++ services/bayn/src/ledger-plan/hash.ts
@@ -1,0 +1,22 @@
+import { Result } from 'effect'
+import type { Account, Transfer } from 'tigerbeetle-node'
+
+import { canonicalHashV1Result } from '../hash'
+import type { LedgerPlan, LedgerPlanHashFailure } from './model'
+
+const serializeRecord = (record: Account | Transfer): Record<string, number | string> =>
+  Object.fromEntries(
+    Object.entries(record).map(([key, value]) => [key, typeof value === 'bigint' ? value.toString() : value]),
+  )
+
+export const hashLedgerPlanResult = (plan: LedgerPlan): Result.Result<string, LedgerPlanHashFailure> =>
+  Result.mapError(
+    canonicalHashV1Result({
+      schemaVersion: 'bayn.ledger-plan.v1',
+      runKey: plan.runKey.toString(),
+      runTag: plan.runTag.toString(),
+      accounts: plan.accounts.map(serializeRecord),
+      transfers: plan.transfers.map(serializeRecord),
+    }),
+    (cause): LedgerPlanHashFailure => ({ _tag: 'LedgerPlanHashCanonicalizationFailed', cause }),
+  )
