+++ services/bayn/src/ledger-plan/input.ts
@@ -1,36 +1,42 @@
-import { Result, Schema } from 'effect'
+import { Result } from 'effect'
 
-import type { EvaluationEvent, InputManifest } from '../types'
+import type { CashYieldEvent, EvaluationEvent, FeeEvent, FillEvent } from '../types'
+import {
+  failLedgerPlan,
+  type LedgerPlanAmountField,
+  type LedgerPlanFailureDetail,
+  type LedgerPlanInputField,
+} from './model'
 
-export interface LedgerInput {
-  readonly runId: string
-  readonly initialCapitalMicros: string
-  readonly inputManifest: InputManifest
-  readonly events: readonly EvaluationEvent[]
+interface DecodedFillEvent {
+  readonly event: FillEvent
+  readonly id: string
+  readonly symbol: string
+  readonly side: FillEvent['side']
+  readonly notionalMicros: bigint
+  readonly costBasisMicros: bigint
 }
 
-export type LedgerPlanInputField =
-  | 'cashYield.amountMicros'
-  | 'cashYield.id'
-  | 'event.kind'
-  | 'events'
-  | 'fee.id'
-  | 'fee.totalMicros'
-  | 'fill.costBasisMicros'
-  | 'fill.id'
-  | 'fill.notionalMicros'
-  | 'fill.side'
-  | 'fill.symbol'
-  | 'initialCapitalMicros'
-  | 'inputManifest.symbol'
-  | 'inputManifest.symbols'
-  | 'runId'
+interface DecodedFeeEvent {
+  readonly event: FeeEvent
+  readonly id: string
+  readonly totalMicros: bigint
+}
 
-export interface LedgerInputDecodeFailure {
-  readonly field: LedgerPlanInputField
-  readonly cause: unknown
-  readonly eventIndex?: number
-  readonly eventKind?: EvaluationEvent['kind']
+interface DecodedCashYieldEvent {
+  readonly event: CashYieldEvent
+  readonly id: string
+  readonly amountMicros: bigint
+}
+
+export interface DecodedLedgerInput {
+  readonly runId: string
+  readonly initialCapitalMicros: bigint
+  readonly inventorySymbols: readonly string[]
+  readonly eventCount: number
+  readonly fills: readonly DecodedFillEvent[]
+  readonly fees: readonly DecodedFeeEvent[]
+  readonly cashYields: readonly DecodedCashYieldEvent[]
 }
 
 interface InputAccessContext {
@@ -39,16 +45,58 @@ interface InputAccessContext {
   eventKind?: EvaluationEvent['kind']
 }
 
-const LedgerInputBoundarySchema = Schema.Struct({
-  runId: Schema.Unknown,
-  initialCapitalMicros: Schema.Unknown,
-  inputManifest: Schema.Struct({
-    symbols: Schema.Array(Schema.Struct({ symbol: Schema.Unknown })),
-  }),
-  events: Schema.Array(Schema.Record(Schema.String, Schema.Unknown)),
+const actualType = (value: unknown): string => (value === null ? 'null' : typeof value)
+
+const inputAccessFailure = (access: InputAccessContext, cause: unknown): LedgerPlanFailureDetail => ({
+  kind: 'input-access-failed',
+  field: access.field,
+  ...(access.eventIndex === undefined ? {} : { eventIndex: access.eventIndex }),
+  ...(access.eventKind === undefined ? {} : { eventKind: access.eventKind }),
+  cause,
 })
 
-const decodeUnknownLedgerMaterial = Schema.decodeUnknownResult(LedgerInputBoundarySchema)
+const invalidInputValue = (
+  field: LedgerPlanInputField,
+  expected: 'evaluation-event-kind' | 'fill-side' | 'string',
+  value: unknown,
+  index?: number,
+  eventKind?: EvaluationEvent['kind'],
+): LedgerPlanFailureDetail => ({
+  kind: 'input-value-invalid',
+  field,
+  expected,
+  actualType: actualType(value),
+  ...(typeof value === 'string' ? { value } : {}),
+  ...(index === undefined ? {} : { index }),
+  ...(eventKind === undefined ? {} : { eventKind }),
+})
+
+const requireString = (
+  field: LedgerPlanInputField,
+  value: unknown,
+  index?: number,
+  eventKind?: EvaluationEvent['kind'],
+): Result.Result<string, LedgerPlanFailureDetail> =>
+  typeof value === 'string'
+    ? Result.succeed(value)
+    : failLedgerPlan(invalidInputValue(field, 'string', value, index, eventKind))
+
+const parseAmount = (
+  field: LedgerPlanAmountField,
+  value: unknown,
+  eventId?: string,
+): Result.Result<bigint, LedgerPlanFailureDetail> =>
+  Result.mapError(
+    Result.try(() => BigInt(value as string)),
+    (cause): LedgerPlanFailureDetail => ({
+      kind: 'amount-parse-failed',
+      field,
+      actualType: actualType(value),
+      ...(typeof value === 'string' ? { value } : {}),
+      ...(eventId === undefined ? {} : { eventId }),
+      cause,
+    }),
+  )
 
 const eventInputField = (kind: unknown, property: PropertyKey): LedgerPlanInputField => {
   if (property === 'kind') return 'event.kind'
@@ -70,64 +118,137 @@ const eventInputField = (kind: unknown, property: PropertyKey): LedgerPlanInputF
   return 'event.kind'
 }
 
-const inputAccessFailure = (access: InputAccessContext, cause: unknown): LedgerInputDecodeFailure => ({
-  field: access.field,
-  ...(access.eventIndex === undefined ? {} : { eventIndex: access.eventIndex }),
-  ...(access.eventKind === undefined ? {} : { eventKind: access.eventKind }),
-  cause,
-})
+const snapshotEvent = (
+  event: unknown,
+  eventIndex: number,
+  access: InputAccessContext,
+): Result.Result<EvaluationEvent, LedgerPlanFailureDetail> => {
+  const inspected = Result.mapError(
+    Result.try(() => {
+      access.field = 'event.kind'
+      access.eventIndex = eventIndex
+      const source = event as Record<PropertyKey, unknown>
+      const kind = source.kind
+      access.eventKind =
+        kind === 'decision' || kind === 'fill' || kind === 'fee' || kind === 'cash-yield' ? kind : undefined
 
-const snapshotEvent = (event: EvaluationEvent, eventIndex: number, access: InputAccessContext): EvaluationEvent => {
-  access.field = 'event.kind'
-  access.eventIndex = eventIndex
-  const kind = event.kind
-  access.eventKind = kind
-
-  const prototype = Object.getPrototypeOf(event)
+      return { source, kind, prototype: Object.getPrototypeOf(source), properties: Reflect.ownKeys(source) }
+    }),
+    (cause) => inputAccessFailure(access, cause),
+  )
+  if (Result.isFailure(inspected)) return Result.fail(inspected.failure)
+  const { source, kind, prototype, properties } = inspected.success
   if (prototype !== Object.prototype && prototype !== null) {
-    throw new TypeError('ledger-plan event must be a plain object')
+    return failLedgerPlan(inputAccessFailure(access, new TypeError('ledger-plan event must be a plain object')))
   }
-
-  const entries = Reflect.ownKeys(event).map((property) => {
+  const entries: [string, unknown][] = []
+  for (const property of properties) {
     access.field = eventInputField(kind, property)
-    if (typeof property !== 'string') throw new TypeError('ledger-plan event must not contain symbol keys')
-    const descriptor = Object.getOwnPropertyDescriptor(event, property)
-    if (descriptor?.enumerable !== true || !('value' in descriptor)) {
-      throw new TypeError('ledger-plan event properties must be enumerable data properties')
+    if (typeof property !== 'string') {
+      return failLedgerPlan(inputAccessFailure(access, new TypeError('ledger-plan event must not contain symbol keys')))
     }
-    return [property, descriptor.value] as const
-  })
-  return Object.fromEntries(entries) as unknown as EvaluationEvent
+    const descriptorResult = Result.mapError(
+      Result.try(() => Object.getOwnPropertyDescriptor(source, property)),
+      (cause) => inputAccessFailure(access, cause),
+    )
+    if (Result.isFailure(descriptorResult)) return Result.fail(descriptorResult.failure)
+    const descriptor = descriptorResult.success
+    if (descriptor?.enumerable !== true || !('value' in descriptor)) {
+      return failLedgerPlan(
+        inputAccessFailure(access, new TypeError('ledger-plan event properties must be enumerable data properties')),
+      )
+    }
+    entries.push([property, descriptor.value])
+  }
+  return Result.succeed(Object.fromEntries(entries) as unknown as EvaluationEvent)
 }
 
-export const decodeLedgerInput = (input: unknown): Result.Result<LedgerInput, LedgerInputDecodeFailure> => {
+const requireEventKind = (
+  event: EvaluationEvent,
+  eventIndex: number,
+): Result.Result<EvaluationEvent['kind'], LedgerPlanFailureDetail> => {
+  const kind = (event as { readonly kind: unknown }).kind
+  return kind === 'decision' || kind === 'fill' || kind === 'fee' || kind === 'cash-yield'
+    ? Result.succeed(kind)
+    : failLedgerPlan(invalidInputValue('event.kind', 'evaluation-event-kind', kind, eventIndex))
+}
+
+const requireFillSide = (
+  value: unknown,
+  eventIndex: number,
+): Result.Result<FillEvent['side'], LedgerPlanFailureDetail> =>
+  value === 'buy' || value === 'sell'
+    ? Result.succeed(value)
+    : failLedgerPlan(invalidInputValue('fill.side', 'fill-side', value, eventIndex, 'fill'))
+
+const decodeEvents = (
+  rawEvents: readonly unknown[],
+  access: InputAccessContext,
+): Result.Result<Pick<DecodedLedgerInput, 'eventCount' | 'fills' | 'fees' | 'cashYields'>, LedgerPlanFailureDetail> =>
+  Result.gen(function* () {
+    const fills: DecodedFillEvent[] = []
+    const fees: DecodedFeeEvent[] = []
+    const cashYields: DecodedCashYieldEvent[] = []
+    for (const [eventIndex, rawEvent] of rawEvents.entries()) {
+      const event = yield* snapshotEvent(rawEvent, eventIndex, access)
+      const kind = yield* requireEventKind(event, eventIndex)
+      if (kind === 'fill') {
+        const fill = event as FillEvent
+        const id = yield* requireString('fill.id', fill.id, eventIndex, kind)
+        const symbol = yield* requireString('fill.symbol', fill.symbol, eventIndex, kind)
+        const side = yield* requireFillSide(fill.side, eventIndex)
+        const notionalMicros = yield* parseAmount('fill.notionalMicros', fill.notionalMicros, id)
+        const costBasisMicros = yield* parseAmount('fill.costBasisMicros', fill.costBasisMicros, id)
+        fills.push({ event: fill, id, symbol, side, notionalMicros, costBasisMicros })
+      } else if (kind === 'fee') {
+        const fee = event as FeeEvent
+        const id = yield* requireString('fee.id', fee.id, eventIndex, kind)
+        const totalMicros = yield* parseAmount('fee.totalMicros', fee.totalMicros, id)
+        fees.push({ event: fee, id, totalMicros })
+      } else if (kind === 'cash-yield') {
+        const cashYield = event as CashYieldEvent
+        const id = yield* requireString('cashYield.id', cashYield.id, eventIndex, kind)
+        const amountMicros = yield* parseAmount('cashYield.amountMicros', cashYield.amountMicros, id)
+        cashYields.push({ event: cashYield, id, amountMicros })
+      }
+    }
+    return { eventCount: rawEvents.length, fills, fees, cashYields }
+  })
+
+export const decodeLedgerInput = (input: unknown): Result.Result<DecodedLedgerInput, LedgerPlanFailureDetail> => {
   let access: InputAccessContext = { field: 'runId' }
   const snapshot = Result.mapError(
     Result.try(() => {
-      const source = input as LedgerInput
+      const source = input as {
+        readonly runId: unknown
+        readonly initialCapitalMicros: unknown
+        readonly inputManifest: { readonly symbols: readonly { readonly symbol: unknown }[] }
+        readonly events: readonly unknown[]
+      }
       access = { field: 'runId' }
-      const runId = source.runId as unknown
+      const runId = source.runId
       access = { field: 'initialCapitalMicros' }
-      const initialCapitalMicros = source.initialCapitalMicros as unknown
+      const initialCapitalMicros = source.initialCapitalMicros
       access = { field: 'inputManifest.symbols' }
       const rawSymbols = [...source.inputManifest.symbols]
       const symbols = rawSymbols.map((coverage, index) => {
         access = { field: 'inputManifest.symbol', eventIndex: index }
-        return { symbol: coverage.symbol as unknown }
+        return coverage.symbol
       })
       access = { field: 'events' }
-      const rawEvents = [...source.events]
-      const events = rawEvents.map((event, eventIndex) => snapshotEvent(event, eventIndex, access))
-      return { runId, initialCapitalMicros, inputManifest: { symbols }, events }
+      return { runId, initialCapitalMicros, symbols, events: [...source.events] }
     }),
     (cause) => inputAccessFailure(access, cause),
   )
   if (Result.isFailure(snapshot)) return Result.fail(snapshot.failure)
 
-  return Result.map(
-    Result.mapError(decodeUnknownLedgerMaterial(snapshot.success), (cause) =>
-      inputAccessFailure({ field: 'runId' }, cause),
-    ),
-    (decoded) => decoded as LedgerInput,
-  )
+  return Result.gen(function* () {
+    const runId = yield* requireString('runId', snapshot.success.runId)
+    const initialCapitalMicros = yield* parseAmount('initialCapitalMicros', snapshot.success.initialCapitalMicros)
+    const inventorySymbols = yield* Result.all(
+      snapshot.success.symbols.map((symbol, index) => requireString('inputManifest.symbol', symbol, index)),
+    )
+    const events = yield* decodeEvents(snapshot.success.events, access)
+    return { runId, initialCapitalMicros, inventorySymbols, ...events }
+  })
 }
